Fix lint

+++ internal/usecase/event_handler.go
@@ -447,7 +447,7 @@ func (eh *eventHandler) processOutbound(event thorchain.Event) error {
 		}
 		if len(unstakeEvt) > 0 && len(unstakeEvt[0].Out) == 2 {
 			err = eh.store.UpdatePoolUnits(models.EventUnstake{
-				StakeUnits: int64(unstakeEvt[0].Events.StakeUnits),
+				StakeUnits: unstakeEvt[0].Events.StakeUnits,
 				Pool:       unstakeEvt[0].Pool,
 				Event: models.Event{
 					Type: unstakeEventType,
